rearranging file structure, rogue server file
