--- helpers/server.js
@@ -1,26 +1,0 @@
-const express = require('express');
-const app = express();
-const MongoClient = require('mongodb').MongoClient;
-const parser = require('body-parser');
-const createRouter = require('./helpers/create_router.js');
-app.use(parser.json());
-
-app.get('/', (req, res) => {
-  console.log("HELLO WORLD");
-});
-
-
-MongoClient.connect('mongodb://localhost:27017')
-.then((client)=> {
-  console.log("HELLO WORLD");
-  const db = client.db('media');
-  const mediaCollection = db.collection('movies');
-  const mediaRouter = createRouter(mediaCollection);
-  console.log(mediaRouter);
-  app.use('/api/movies', mediaRouter);
-})
-.catch(console.err);
-
-app.listen(5000, function(){
-  console.log(`Listening on port ${ this.address().port}`);
-});
