+++ server.js
@@ -1,19 +1,26 @@
-const path = require('path');
-const cors = require('cors');
 const express = require('express');
 const app = express();
-const bodyParser = require('body-parser');
-
-app.use(bodyParser.json());
-app.use(cors());
 const MongoClient = require('mongodb').MongoClient;
+const parser = require('body-parser');
 const createRouter = require('./helpers/create_router.js');
+app.use(parser.json());
 
-app.get('/api', (req, res) => {
-  console.log("INDEX route has been requested");
-  res.json({"message": "Hello world"});
+app.get('/', (req, res) => {
+  console.log("HELLO WORLD");
 });
 
-app.listen(5000, function() {
-  console.log(`All the magic is happening on port ${this.address().port}`);
+
+MongoClient.connect('mongodb://localhost:27017')
+.then((client)=> {
+  console.log("HELLO WORLD");
+  const db = client.db('media');
+  const mediaCollection = db.collection('movies');
+  const mediaRouter = createRouter(mediaCollection);
+  console.log(mediaRouter);
+  app.use('/api/movies', mediaRouter);
+})
+.catch(console.err);
+
+app.listen(5000, function(){
+  console.log(`Listening on port ${ this.address().port}`);
 });
